+++ src/main/scala/scorex/core/utils/SerializationConstants.scala
@@ -1,0 +1,6 @@
+package scorex.core.utils
+
+object SerializationConstants {
+  val IntSize = 4
+  val LongSize = 8
+}